[pf-439] Renamed function based files with dash-case
