+++ api/lib/domain/strategies/challenges-filter.js
